+++ include/ignition/sensors/Lidar.hh
@@ -77,6 +77,11 @@ namespace ignition
       /// \return true if the update was successfull
       public: virtual bool PublishLidarScan(const common::Time &_now);
 
+      /// \brief Load the sensor based on data from an sdf::Sensor object.
+      /// \param[in] _sdf SDF Sensor parameters.
+      /// \return true if loading was successful
+      public: virtual bool Load(const sdf::Sensor &_sdf) override;
+
       /// \brief Load the sensor with SDF parameters.
       /// \param[in] _sdf SDF Sensor parameters.
       /// \return true if loading was successful
@@ -135,19 +140,19 @@ namespace ignition
 
       /// \brief Get the ray count
       /// \return The number of rays
-      public: int RayCount() const;
+      public: unsigned int RayCount() const;
 
       /// \brief Get the range count
       /// \return The number of ranges
-      public: int RangeCount() const;
+      public: unsigned int RangeCount() const;
 
       /// \brief Get the vertical scan line count
       /// \return The number of scan lines vertically
-      public: int VerticalRayCount() const;
+      public: unsigned int VerticalRayCount() const;
 
       /// \brief Get the vertical scan line count
       /// \return The number of scan lines vertically
-      public: int VerticalRangeCount() const;
+      public: unsigned int VerticalRangeCount() const;
 
       /// \brief Get the vertical scan bottom angle
       /// \return The minimum angle of the scan block