+++ src/GaussianNoiseModel.cc
@@ -47,6 +47,14 @@ class ignition::sensors::GaussianNoiseModelPrivate
   /// \brief If type starts with GAUSSIAN, the bias we'll add.
   public: double bias = 0.0;
 
+  /// \brief If type starts with GAUSSIAN, the standard deviation of the
+  /// distribution from which the dynamic bias will be driven.
+  public: double dynamicBiasStdDev = 0.0;
+
+  /// \brief If type starts with GAUSSIAN, the correlation time of the
+  /// process from which the dynamic bias will be driven.
+  public: double dynamicBiasCorrTime = 0.0;
+
   /// \brief If type==GAUSSIAN_QUANTIZED, the precision to which
   /// the output signal is rounded.
   public: double precision = 0.0;
@@ -90,6 +98,8 @@ void GaussianNoiseModel::Load(const sdf::Noise &_sdf)
 
   this->dataPtr->mean = _sdf.Mean();
   this->dataPtr->stdDev = _sdf.StdDev();
+  this->dataPtr->dynamicBiasStdDev = _sdf.DynamicBiasStdDev();
+  this->dataPtr->dynamicBiasCorrTime = _sdf.DynamicBiasCorrelationTime();
 
   // Sample the bias
   double biasMean = 0;
@@ -114,11 +124,33 @@ void GaussianNoiseModel::Load(const sdf::Noise &_sdf)
 }
 
 //////////////////////////////////////////////////
-double GaussianNoiseModel::ApplyImpl(double _in)
+double GaussianNoiseModel::ApplyImpl(double _in, double _dt)
 {
-  // Add independent (uncorrelated) Gaussian noise to each input value.
+  // Generate independent (uncorrelated) Gaussian noise to each input value.
   double whiteNoise = ignition::math::Rand::DblNormal(
       this->dataPtr->mean, this->dataPtr->stdDev);
+
+  // Generate varying (correlated) bias to each input value.
+  // This implementation is based on the one available in Rotors:
+  // https://github.com/ethz-asl/rotors_simulator/blob/master/rotors_gazebo_plugins/src/gazebo_imu_plugin.cpp
+  //
+  // More information about the parameters and their derivation:
+  //
+  // https://github.com/ethz-asl/kalibr/wiki/IMU-Noise-Model
+
+  if (this->dataPtr->dynamicBiasStdDev > 0 &&
+     this->dataPtr->dynamicBiasCorrTime > 0)
+  {
+    double sigma_b = this->dataPtr->dynamicBiasStdDev;
+    double tau = this->dataPtr->dynamicBiasCorrTime;
+
+    double sigma_b_d = sqrt(-sigma_b * sigma_b *
+        tau / 2 * expm1(-2 * _dt / tau));
+    double phi_d = exp(-_dt / tau);
+    this->dataPtr->bias = phi_d * this->dataPtr->bias +
+      ignition::math::Rand::DblNormal(0, sigma_b_d);
+  }
+
   double output = _in + this->dataPtr->bias + whiteNoise;
 
   if (this->dataPtr->quantized)