+++ Changelog.md
@@ -2,6 +2,12 @@
 
 ### Ignition Sensors 2.X.X (201X-XX-XX)
 
+1. Add support for loading a Lidar sensor from an SDF Sensor DOM object.
+    * [Pull request 59](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/59)
+
+1. Add support for loading an IMU sensor from an SDF Sensor DOM object.
+    * [Pull request 58](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/58)
+
 1. Add support for loading a camera and depth camera sensor from an SDF Sensor DOM object.
     * [Pull request 57](https://bitbucket.org/ignitionrobotics/ign-sensors/pull-request/57)